:recycle: refactor: handleSelect 훅으로 빼기
